LayoutTests/accessibility/mac/text-input-marked-range.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test.js"></script>
</head>
<body>
    <input id="editable" type="text"></input>
    <script type="text/javascript">
        description("This test ensures that the input method marked range is available to accessibility clients.");
        editable.focus();
        if (window.textInputController && window.accessibilityController) {
            var element = accessibilityController.accessibleElementById("editable");
            shouldBeTrue("!element.textInputMarkedRange");
            
            textInputController.setMarkedText("test", 1, 0);
            shouldBe("element.textInputMarkedRange", "'{0, 4}'");
            
            textInputController.insertText("t");
            shouldBeTrue("!element.textInputMarkedRange");
            
            textInputController.setMarkedText("test2", 1, 0);
            shouldBe("element.textInputMarkedRange", "'{1, 5}'");
        }
    </script>
</body>
</html>
